+++ src/ExportManager/OpenSheet.cs
@@ -56,7 +56,7 @@ namespace SCaddins.ExportManager
             collector.OfCategory(BuiltInCategory.OST_Sheets);
             foreach (var element in collector) {
                 var view = (ViewSheet)element;
-                #if REVIT2019 || REVIT2020 || REVIT2021
+                #if REVIT2019 || REVIT2020 || REVIT2021 || REVIT2022
                 result.Add(new OpenableView(view.Name, view.SheetNumber, view));
                 #else
                 result.Add(new OpenableView(view.ViewName, view.SheetNumber, view));