+++ src/LineOfSight/StaduimSeatingTier.cs
@@ -1,4 +1,4 @@
-﻿// (C) Copyright 2012-2020 by Andrew Nicholas
+﻿// (C) Copyright 2012-2021 by Andrew Nicholas
 //
 // This file is part of SCaddins.
 //
@@ -233,7 +233,7 @@ namespace SCaddins.LineOfSight
                 return null;
             }
             var viewDrafting = ViewDrafting.Create(doc, viewFamilyType.Id);
-            #if REVIT2019 || REVIT2020 || REVIT2021
+            #if REVIT2019 || REVIT2020 || REVIT2021 || REVIT2022
             viewDrafting.Name = newViewName;
             #else
             view.ViewName = newViewName;