+++ src/SolarAnalysis/SolarAnalysisCommand.cs
@@ -1,4 +1,4 @@
-// (C) Copyright 2013-2020 by Andrew Nicholas
+// (C) Copyright 2013-2021 by Andrew Nicholas
 //
 // This file is part of SCaddins.
 //
@@ -46,7 +46,7 @@ namespace SCaddins.SolarAnalysis
             SCaddinsApp.WindowManager.ShowDialog(vm, null, ViewModels.SolarViewsViewModel.DefaultViewSettings);
 
             if (vm.CreateAnalysisView) {
-                #if REVIT2021
+                #if REVIT2021 || REVIT2022
                 var internalUnitsGridSize = UnitUtils.ConvertToInternalUnits(vm.AnalysisGridSize, UnitTypeId.Millimeters);
                 #else
                 var internalUnitsGridSize = UnitUtils.ConvertToInternalUnits(vm.AnalysisGridSize, DisplayUnitType.DUT_MILLIMETERS);